[Jeet] Refactored the utils.py and moved inventory call out of it to fabric_inventory.py
